feat: 사용자 데이터를 로딩하는 동안 스켈레톤 로더를 추가
